refactor(api): move AdminMemberError into team context
